地方撒旦是

+++ station-bundle-client/src/main/java/com/taobao/cun/auge/station/service/PartnerInstanceService.java
@@ -65,7 +65,7 @@ public interface PartnerInstanceService {
 	 * @param frozenMoney
 	 * @return
 	 */
-	public boolean freezeBond(Long taobaoUserId, BigDecimal frozenMoney) throws AugeServiceException;
+	public boolean freezeBond(Long taobaoUserId, Double frozenMoney) throws AugeServiceException;
 
 	/**
 	 * 开业
